requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
defusedxml==0.7.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.8
fonttools==4.58.4
fpdf2==2.8.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.6
Mako==1.3.8
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
passlib==1.7.4
pillow==11.2.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
